Update README.md
There are three files that are uploaded, including one dataset (CSV.file)downloaded from Kaggle, one rmd file of the code for the model, and a pdf file about the report of the model. If you want to run my code from rmd file, please download the CSV file, and change the path to yours. 

In the report, it would introduce the study about the youth's alcohol consumption level, including abstract, introduction, dataset description, model explanation, and discussion about the results.